Organize the game more.
