+++ minesweeper/__main__.py
@@ -1,0 +1,5 @@
+from .minesweeper import Minesweeper
+
+
+if __name__ == "__main__":
+    Minesweeper().start()
